Move JSON parsing/file writing out of WhippetLock into abstract class
