+++ src/Files/WhippetLock.php
@@ -2,23 +2,8 @@
 
 namespace Dxw\Whippet\Files;
 
-class WhippetLock
+class WhippetLock extends Base
 {
-    public static function fromString(/* string */ $json)
-    {
-        return new self(json_decode($json, true));
-    }
-
-    public static function fromFile(/* string */ $path)
-    {
-        return self::fromString(file_get_contents($path));
-    }
-
-    public function __construct(array $data)
-    {
-        $this->data = $data;
-    }
-
     public function getDependencies(/* string */ $type)
     {
         if (!isset($this->data[$type])) {
@@ -45,10 +30,5 @@ class WhippetLock
             'src' => $src,
             'revision' => $revision,
         ];
-    }
-
-    public function saveToPath(/* string */ $path)
-    {
-        file_put_contents($path, json_encode($this->data, JSON_PRETTY_PRINT | JSON_UNESCAPED_SLASHES)."\n");
     }
 }
